Move dummy objects, encoding and decoding into seperate file
