+++ src/DataReader/mod.rs
@@ -3,4 +3,7 @@ pub mod json_writer;
 pub mod parse_edge;
 pub mod parse_invariant;
 pub mod parse_queries;
+pub mod serialization;
 pub mod xml_parser;
+
+use serialization::{DummyComponent, DummyEdge, DummyLocation};
